Add TC2 preparations
This commit updates the prisma schema to include tournament information. It also moves commonly used utility functions to ServerUtils.js for future use. This commit also creates an algorithm.js file which will be used for the tournament scheduling algorithm. The TC1 imports are updated as needed.
