+++ backend/tournamentScheduling/algorithm.js
@@ -1,0 +1,1 @@
+//TODO: Tournament scheduling algorithm that matches teams on the platform to play against each other
